+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added attributes
+
+* `hx_confirm`
+
 
 ## [1.0.0] - 2024-11-03